Sql handling in tests fixed due to id generation policy changes (BigInteger fields are selected not as Long but as BigDecimal)

+++ src/main/java/com/anli/busstation/dal/sql/test/IdSelector.java
@@ -2,6 +2,7 @@ package com.anli.busstation.dal.sql.test;
 
 import com.anli.sqlexecution.handling.ResultSetHandler;
 import com.anli.sqlexecution.handling.TransformingResultSet;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -13,7 +14,7 @@ public class IdSelector implements ResultSetHandler<List<BigInteger>> {
     public List<BigInteger> handle(TransformingResultSet resultSet) throws SQLException {
         List<BigInteger> idList = new ArrayList<>();
         while (resultSet.next()) {
-            idList.add(BigInteger.valueOf(resultSet.getValue(1, Long.class)));
+            idList.add(resultSet.getValue(1, BigDecimal.class).toBigInteger());
         }
         return idList;
     }
